Внесены исправления кнопок, добавлены скругление картинки, добавлено исчезновение рамки после ответа. Статус бар изменен на белый, модели вынесены в отдельные файлы. К кнопкам добавлено выключение, после выбора правильного ответа, и до появления следующего вопроса.
