+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Дмитрий on 28.01.2024.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
