+++ MovieQuiz/Resources/Info.plist
@@ -2,11 +2,11 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-    <key>UIAppFonts</key>
-    <array>
-        <string>YS Display-Bold.ttf</string>
-        <string>YS Display-Medium.ttf</string>
-    </array>
+	<key>UIAppFonts</key>
+	<array>
+		<string>YS Display-Bold.ttf</string>
+		<string>YS Display-Medium.ttf</string>
+	</array>
 	<key>UIApplicationSceneManifest</key>
 	<dict>
 		<key>UIApplicationSupportsMultipleScenes</key>
